# Laravel/projecte_final/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example for response examples value",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/Projecte_Final/Laravel/projecte_final/public/index.php",
            "description": "L5 Swagger OpenApi Server"
        }
    ],
    "paths": {
...
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Veure totes les clients.",
                "operationId": "0a140f3fb03571bf1c4e5e9360eab526",
                "responses": {
                    "200": {
                        "description": "Retorna totes les clients."
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/client/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Veure totes las clients.",
                "operationId": "06fa1a3b3beb5ae40b72059694a1c8ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna totes les clients."
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Update client",
                "operationId": "2e8866ad96c2cc37286cbde160c1d8f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nom",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user"
                    },
                    {
                        "name": "cognoms",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user"
                    },
                    {
                        "name": "nif",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "12345"
                    },
                    {
                        "name": "tipu_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Delete client",
                "operationId": "73369112a57e6310cd8bcddffcba9a84",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Cita not found"
                    }
                }
            }
        },
        "/api/client": {
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Adds a new client",
                "operationId": "9f8160de9035b77d036ad5ad32645606",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nom",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user"
                    },
                    {
                        "name": "cognoms",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user"
                    },
                    {
                        "name": "nif",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "12345"
                    },
                    {
                        "name": "tipu_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/facturas": {
            "get": {
                "tags": [
                    "Facturas"
                ],
                "summary": "Veure totes les factures.",
                "operationId": "3f2ca30a5d394f811c0aed55879ef64c",
                "responses": {
                    "200": {
                        "description": "Retorna totes les factures."
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/factura": {
            "post": {
                "tags": [
                    "Facturas"
                ],
                "summary": "Adds a new factura",
                "operationId": "e683ef93d57db7297b0639cb84ebb20d",
                "parameters": [
                    {
                        "name": "data",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2023-07-20"
                    },
                    {
                        "name": "total",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "100"
                    },
                    {
                        "name": "total_con_iva",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "121"
                    },
                    {
                        "name": "client_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vehicle_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "serveis[0]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "serveis[1]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}